x509-cert: move CSR builder to x509_cert::request
